added feature to bypass unsupported image
it will post a default picture since I couldn't get PIL or Pillow to work on my PC to resize the image. Hopefully I would be able to fix it later on. for now, it will post a default image with all the other info.

Also for some reason, the publish_media response is sometimes 403: Forbidden, but the image is still posted successfully
will look into it later